pyproject.toml
[tool.poetry]
name = "django-example"
version = "0.1.0"
description = ""
authors = ["Ainur Vakhitov <ainur.vahitov2002@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
django-ninja = "^1.3.0"
punq = "^0.7.0"
pillow = "^11.0.0"
yagmail = "^0.15.293"
celery = {extras = ["redis"], version = "^5.4.0"}
gevent = "^24.11.1"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
pre-commit = "^4.0.1"
faker = "^30.8.2"
pytest = "^8.3.3"
pytest-django = "^4.9.0"
factory-boy = "^3.3.1"
autoflake = "^2.3.1"
kolo = "^2.34.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings.local"

[tool.autoflake]
imports = ["django", "requests", "urllib3"]
in-place=true
remove-all-unused-imports=true
ignore-init-module-imports=true
remove-unused-variables=true

[tool.isort]
multi_line_output=3
include_trailing_comma=true
line_length=120
lines_after_imports=2
lines_between_sections=1
lines_between_types=0
force_alphabetical_sort_within_sections=true
ensure_newline_before_comments=true
combine_as_imports=true
combine_star=true
skip_gitignore=true
balanced_wrapping=true
force_grid_wrap=2
src_paths=["core", ]
skip_glob=[
    ".git",
    ".idea",
    ".vscode",
    "venv",
    ".venv",
]
known_django=["django","ninja"]
known_first_party=["core"]
sections=[
    "FUTURE",
    "STDLIB",
    "DJANGO",
    "THIRDPARTY",
    "FIRSTPARTY",
    "LOCALFOLDER",
]